+++ src/Cwaff-Guns/Bouncer.cs
@@ -16,7 +16,7 @@ public class Bouncer : CwaffGun
     {
         Lazy.SetupGun<Bouncer>(ItemName, ShortDescription, LongDescription, Lore)
           .SetAttributes(quality: ItemQuality.D, gunClass: GunClass.PISTOL, reloadTime: 1.3f, ammo: 300, shootFps: 14, reloadFps: 30,
-            muzzleFrom: Items.Magnum, fireAudio: "MC_RocsCape")
+            muzzleFrom: Items.Magnum, fireAudio: "MC_RocsCape", smoothReload: 0.1f)
           .SetReloadAudio("bouncer_reload_short", 5, 10, 15, 20)
           .SetReloadAudio("bouncer_reload", 25)
           .AddToShop(ModdedShopType.Boomhildr)