+++ src/Cwaff-Guns/HatchlingGun.cs
@@ -11,7 +11,7 @@ public class HatchlingGun : CwaffGun
     {
         Lazy.SetupGun<HatchlingGun>(ItemName, ShortDescription, LongDescription, Lore)
           .SetAttributes(quality: ItemQuality.D, gunClass: GunClass.RIFLE, reloadTime: 1.25f, ammo: 500,
-            shootFps: 40, reloadFps: 20, muzzleFrom: Items.Mailbox)
+            shootFps: 40, reloadFps: 20, muzzleFrom: Items.Mailbox, smoothReload: 0.1f)
           .SetReloadAudio("hatchling_gun_bounce_sound", 0, 6, 14)
           .InitProjectile(GunData.New(clipSize: 12, cooldown: 0.2f, angleVariance: 15.0f, shootStyle: ShootStyle.SemiAutomatic, customClip: true,
             damage: 3.0f, speed: 24.0f, sprite: "egg", fps: 12, scale: 1.5f, anchor: Anchor.MiddleCenter, spawnSound: "hatchling_gun_shoot_sound"))