+++ src/Cwaff-Guns/Nycterian.cs
@@ -13,7 +13,7 @@ public class Nycterian : CwaffGun
     {
         Lazy.SetupGun<Nycterian>(ItemName, ShortDescription, LongDescription, Lore)
           .SetAttributes(quality: ItemQuality.C, gunClass: GunClass.PISTOL, reloadTime: 1.1f, ammo: 325, shootFps: 20, reloadFps: 20,
-            muzzleFrom: Items.Mailbox, fireAudio: "nycterian_shoot_sound", reloadAudio: "nycterian_reload_sound")
+            muzzleFrom: Items.Mailbox, fireAudio: "nycterian_shoot_sound", reloadAudio: "nycterian_reload_sound", smoothReload: 0.1f)
           .InitProjectile(GunData.New(clipSize: 10, cooldown: 0.3f, shootStyle: ShootStyle.SemiAutomatic, scale: 1.5f, customClip: true,
             damage: 7.0f, speed: 18.0f, range: 100.0f, sprite: "bat_projectile", fps: 12))
           .Attach<BounceProjModifier>(bounce => { bounce.numberOfBounces = 1; })