+++ src/Cwaff-Guns/Widowmaker.cs
@@ -16,7 +16,8 @@ public class Widowmaker : CwaffGun
     public static void Init()
     {
         Lazy.SetupGun<Widowmaker>(ItemName, ShortDescription, LongDescription, Lore)
-          .SetAttributes(quality: ItemQuality.B, gunClass: GunClass.RIFLE, reloadTime: 1.4f, ammo: 160, shootFps: 20, reloadFps: 12, fireAudio: "widowmaker_fire_sound")
+          .SetAttributes(quality: ItemQuality.B, gunClass: GunClass.RIFLE, reloadTime: 1.4f, ammo: 160, shootFps: 20, reloadFps: 12,
+            fireAudio: "widowmaker_fire_sound", smoothReload: 0.1f)
           .SetReloadAudio("widowmaker_reload_sound", 0, 4, 8, 10, 12, 14)
           .InitProjectile(GunData.New(clipSize: 5, cooldown: 0.18f, shootStyle: ShootStyle.SemiAutomatic, damage: 3.5f, pierceBreakables: true,
             sprite: "widowmaker_projectile", fps: 12, scale: _SCALE, anchor: Anchor.MiddleLeft, preventOrbiting: true, customClip: true))