+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.22.4 (TBD)
+
+- Added dynamic reload animation speeds to several guns (i.e., their animations reflect changes to the player's reload speed stat)
+	- (Bouncer, Bubblebeam, Carpet Bomber, Derail Gun, Flakseed, Hand Cannon, Hatchling Gun, Lightwing, Nycterian, Platinum Star, Sub Machine Gun, Subtractor Beam, Tranquilizer, Uppskeruvel, Wavefront, and Widowmaker)
+
 ## 1.22.3 (2025-03-21)
 
 #### Features: